added full functionality. Pending testing of this code, it *might* actually open the safe
Read long paragraph of comments at the very top of the program. It has some notes on things that should be tested in case the code doesn't initially work. Try these things first. These are mostly library-related things that are likely to cause problems if the code doesn't work. This code has been modified to attempt to crack the safe using the AccelStepper.h library. Note that a kill switch should be added to mitigate stalling after the safe opens!